address the rest of PR comments
