+++ integrations/cleverAWS/aws.go
@@ -18,16 +18,13 @@ var (
 // AwsService does the computation to form AWS usernames with a first initial and last name.
 type AwsService struct{}
 
-// Init implements the InfoSource interface. There are no API calls made.
-func (a AwsService) Init(_ string) error { return nil }
-
 // Fill uses the first and last name to form an AWS username.
-func (a AwsService) Fill(m integrations.UserMap) integrations.UserMap {
+func (a AwsService) Fill(m integrations.UserMap) (integrations.UserMap, error) {
 	for email, user := range m {
 		if user.FirstName != "" && user.LastName != "" {
 			user.AWS = strings.ToLower(user.FirstName[0:1] + user.LastName)
 		}
 		m[email] = user
 	}
-	return m
+	return m, nil
 }
