+++ main.go
@@ -1,16 +1,13 @@
 package main
 
 import (
-	"encoding/json"
 	"fmt"
 	"log"
 	"net/http"
 	"os"
 
+	"github.com/Clever/who-is-who/api"
 	"github.com/Clever/who-is-who/integrations"
-	"github.com/Clever/who-is-who/integrations/cleverAWS"
-	"github.com/Clever/who-is-who/integrations/slack"
-	"github.com/gorilla/mux"
 	"gopkg.in/clever/kayvee-go.v2"
 )
 
@@ -19,8 +16,8 @@ var (
 	awsKey         string
 	awsSecret      string
 	dynamoTable    string
-	dynamoEndpoint string
 	dynamoRegion   string
+	dynamoEndpoint string
 )
 
 // m is a convenience type for using kayvee.
@@ -43,83 +40,8 @@ func setupEnvVars() {
 	awsKey = requiredEnv("AWS_ACCESS_KEY")
 	awsSecret = requiredEnv("AWS_SECRET_KEY")
 	dynamoTable = requiredEnv("DYNAMO_TABLE")
-	dynamoEndpoint = requiredEnv("DYNAMO_ENDPOINT")
 	dynamoRegion = requiredEnv("DYNAMO_REGION")
-}
-
-// dynamoConn wraps the Dynamo client and it's helper functions for Users.
-type dynamoConn struct {
-	Dynamo integrations.Client
-}
-
-// aliasEndpoint generates a HTTP handler that will query for a user with the specified
-// dynamo index and provided value.
-func (d dynamoConn) listEndpoint(w http.ResponseWriter, r *http.Request) {
-	// query Dynamo for all users
-	users, err := d.Dynamo.GetUserList()
-	if err != nil {
-		log.Println(kayvee.FormatLog("who-is-who", kayvee.Error, "getUserList error", m{
-			"message": err.Error(),
-		}))
-		w.WriteHeader(http.StatusInternalServerError)
-		return
-	}
-
-	// write the user to the connection
-	err = json.NewEncoder(w).Encode(users)
-	if err != nil {
-		log.Println(kayvee.FormatLog("who-is-who", kayvee.Error, "json encoding error", m{
-			"message": err.Error(),
-		}))
-		w.WriteHeader(http.StatusInternalServerError)
-		return
-	}
-}
-
-// aliasEndpoint generates a HTTP handler that will query for a user with the specified
-// dynamo index and provided value.
-func (d dynamoConn) aliasEndpoint(idx integrations.Index, key string) func(w http.ResponseWriter, r *http.Request) {
-	return func(w http.ResponseWriter, r *http.Request) {
-		// get the username/alias from the URL parameter
-		username, valid := mux.Vars(r)[key]
-		if !valid {
-			w.WriteHeader(http.StatusBadRequest)
-			return
-		}
-
-		// query Dynamo for the user with the specified details.
-		user, err := d.Dynamo.GetUser(idx, username)
-		if err == integrations.ErrUserDNE {
-			w.WriteHeader(http.StatusNotFound)
-			return
-		} else if err != nil {
-			log.Println(kayvee.FormatLog("who-is-who", kayvee.Error, "getUser error", m{
-				"message": err.Error(),
-			}))
-			w.WriteHeader(http.StatusInternalServerError)
-			return
-		}
-
-		// write the user to the connection
-		err = json.NewEncoder(w).Encode(user)
-		if err != nil {
-			log.Println(kayvee.FormatLog("who-is-who", kayvee.Error, "json encoding error", m{
-				"message": err.Error(),
-			}))
-			w.WriteHeader(http.StatusInternalServerError)
-			return
-		}
-	}
-}
-
-// hookUpRouter sets up the router
-func hookUpRouter(d dynamoConn) *mux.Router {
-	r := mux.NewRouter()
-	r.HandleFunc("/list", d.listEndpoint)
-	r.HandleFunc("/alias/aws/{username}", d.aliasEndpoint(cleveraws.Index, "username"))
-	r.HandleFunc("/alias/slack/{handle}", d.aliasEndpoint(slack.Index, "handle"))
-	r.HandleFunc("/alias/email/{email}", d.aliasEndpoint(integrations.EmailIndex, "email"))
-	return r
+	dynamoEndpoint = requiredEnv("DYNAMO_ENDPOINT")
 }
 
 func main() {
@@ -132,7 +54,7 @@ func main() {
 			"message": err.Error(),
 		}))
 	}
-	d := dynamoConn{
+	d := api.DynamoConn{
 		Dynamo: c,
 	}
 
@@ -140,5 +62,5 @@ func main() {
 	log.Println(kayvee.FormatLog("who-is-who", kayvee.Info, "server startup", m{
 		"message": fmt.Sprintf("Listening on %s", port),
 	}))
-	http.ListenAndServe(port, hookUpRouter(d))
+	http.ListenAndServe(port, d.HookUpRouter())
 }
